v3.5+ add sequence increment and update dependency version
This commit will only work for Bitfocus version 3.5 and above
due to the runtime update from node 18 to node 22.

Added a sequence increment to the packet header. This will allow
quicker responses for successive button pushes and conform more
with the Visca protocol.

Updated the dependency and yarn versions in package.json.

Updated the runtime node version to 22 in manifest.json.

Added .yarnrc.yml as it was needed for the newer yarn version.